Restructure code into correct files
