--- homeassistant/components/geocaching/entities.py
@@ -1,391 +1,0 @@
-"""Entities for Geocaching devices."""
-
-from collections.abc import Callable
-from dataclasses import dataclass
-import datetime
-from typing import Any, cast
-
-from geocachingapi.models import (
-    GeocachingCache,
-    GeocachingTrackable,
-    GeocachingTrackableJourney,
-)
-
-from homeassistant.components.device_tracker.config_entry import TrackerEntity
-from homeassistant.components.sensor import (
-    SensorDeviceClass,
-    SensorEntity,
-    SensorEntityDescription,
-)
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
-from homeassistant.helpers.entity import Entity
-from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
-
-from .const import (
-    CACHE_ID_SENSOR_FORMAT,
-    DOMAIN,
-    TRACKABLE_ID_SENSOR_FORMAT,
-    GeocacheCategory,
-)
-from .coordinator import GeocachingDataUpdateCoordinator
-
-
-# A device can have multiple entities, and for a cache which requires multiple entities we want to group them together.
-# Therefore, we create a device for each cache, which holds all related entities.
-# This function returns the device info for a cache.
-def get_cache_device_info(
-    cache: GeocachingCache, category: GeocacheCategory
-) -> DeviceInfo:
-    """Generate device info for a cache."""
-    return DeviceInfo(
-        name=f"Geocache {category.value} {cache.reference_code}",
-        identifiers={(DOMAIN, cast(str, cache.reference_code))},
-        entry_type=DeviceEntryType.SERVICE,
-        manufacturer="Groundspeak, Inc.",
-    )
-
-
-# A device can have multiple entities, and for a trackable which requires multiple entities we want to group them together.
-# Therefore, we create a device for each trackable, which holds all related entities.
-# This function returns the device info for a trackable.
-def get_trackable_device_info(trackable: GeocachingTrackable) -> DeviceInfo:
-    """Generate device info for a cache."""
-    return DeviceInfo(
-        name=f"Geotrackable {trackable.reference_code}",
-        identifiers={(DOMAIN, cast(str, trackable.reference_code))},
-        entry_type=DeviceEntryType.SERVICE,
-        manufacturer="Groundspeak, Inc.",
-    )
-
-
-# pylint: disable=hass-enforce-class-module
-# Base class for a cache entity.
-# Sets the device, ID and translation settings to correctly group the entity to the correct cache device and give it the correct name.
-class GeoEntityBaseCache(CoordinatorEntity[GeocachingDataUpdateCoordinator], Entity):
-    """Base class for cache entities."""
-
-    _attr_has_entity_name = True
-    cache: GeocachingCache
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        cache: GeocachingCache,
-        entity_type: str,
-        category: GeocacheCategory,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator)
-        self.cache = cache
-
-        # Set the device info from the cache, to group all entities for a cache together
-        self._attr_device_info = get_cache_device_info(cache, category)
-
-        self._attr_unique_id = CACHE_ID_SENSOR_FORMAT.format(
-            category.value, cache.reference_code, entity_type
-        )
-
-        # The translation key determines the name of the entity as this is the lookup for the `strings.json` file.
-        self._attr_translation_key = f"cache_{entity_type}"
-
-
-# pylint: disable=hass-enforce-class-module
-# Base class for a trackable entity.
-# Sets the device, ID and translation settings to correctly group the entity to the correct trackable device and give it the correct name.
-class GeoEntityBaseTrackable(
-    CoordinatorEntity[GeocachingDataUpdateCoordinator], Entity
-):
-    """Base class for trackable entities."""
-
-    _attr_has_entity_name = True
-    trackable: GeocachingTrackable
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        trackable: GeocachingTrackable,
-        entity_type: str,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator)
-        self.trackable = trackable
-
-        # Set the device info from the trackable, to group all entities for a trackable together
-        self._attr_device_info = get_trackable_device_info(trackable)
-
-        self._attr_unique_id = TRACKABLE_ID_SENSOR_FORMAT.format(
-            trackable.reference_code, entity_type
-        )
-
-        # The translation key determines the name of the entity as this is the lookup for the `strings.json` file.
-        self._attr_translation_key = f"trackable_{entity_type}"
-
-
-# pylint: disable=hass-enforce-class-module
-# A tracker entity that allows us to show caches on a map.
-class GeoEntityCacheLocation(GeoEntityBaseCache, TrackerEntity):
-    """Entity for a cache GPS location."""
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        cache: GeocachingCache,
-        category: GeocacheCategory,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator, cache, "location", category)
-
-    @property
-    def native_value(self) -> str | None:
-        """Return the reference code of the cache."""
-        return self.cache.reference_code
-
-    @property
-    def latitude(self) -> float:
-        """Return the latitude of the cache."""
-        return float(self.cache.coordinates.latitude or 0)
-
-    @property
-    def longitude(self) -> float:
-        """Return the longitude of the cache."""
-        return float(self.cache.coordinates.longitude or 0)
-
-    @property
-    def location_name(self) -> str | None:
-        """Return the location of the cache."""
-        # TODO: Figure out another way of displaying the code as label in the map, rather than this with `label_mode: state` in the map | pylint: disable=fixme
-        return self.cache.reference_code
-
-
-# pylint: disable=hass-enforce-class-module
-# A tracker entity that allows us to show caches on a map.
-class GeoEntityTrackableLocation(GeoEntityBaseTrackable, TrackerEntity):
-    """Entity for a trackable GPS location."""
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        trackable: GeocachingTrackable,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator, trackable, "location")
-
-    @property
-    def native_value(self) -> str | None:
-        """Return the reference code of the trackable."""
-        return self.trackable.reference_code
-
-    @property
-    def latitude(self) -> float:
-        """Return the latitude of the trackable."""
-        return float(self.trackable.coordinates.latitude or 0)
-
-    @property
-    def longitude(self) -> float:
-        """Return the longitude of the trackable."""
-        return float(self.trackable.coordinates.longitude or 0)
-
-    @property
-    def location_name(self) -> str | None:
-        """Return the location of the cache."""
-        # TODO: Figure out another way of displaying the code as label in the map, rather than this with `label_mode: state` in the map | pylint: disable=fixme
-        return self.trackable.reference_code
-
-    def _format_travel_log_entry(
-        self, journey: GeocachingTrackableJourney
-    ) -> dict[str, Any]:
-        """Format a single journey entry."""
-        distance: str = (
-            "Unknown"
-            if journey.distance_km is None
-            else f"{round(journey.distance_km)} km"
-        )
-        return {
-            "date": journey.date,
-            "username": journey.user.username,
-            "location_name": journey.location_name,
-            "distance_travelled": distance,
-        }
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any] | None:
-        """Return the state attributes."""
-        return {
-            "travel_log": []
-            if self.trackable.journeys is None
-            else [
-                self._format_travel_log_entry(journey)
-                for journey in self.trackable.journeys
-            ],
-        }
-
-
-@dataclass(frozen=True, kw_only=True)
-class GeocachingCacheSensorDescription(SensorEntityDescription):
-    """Define Sensor entity description class."""
-
-    value_fn: Callable[[GeocachingCache], StateType | datetime.date]
-
-
-@dataclass(frozen=True, kw_only=True)
-class GeocachingTrackableEntityDescription(SensorEntityDescription):
-    """Define Sensor entity description class."""
-
-    value_fn: Callable[[GeocachingTrackable], StateType | datetime.date]
-
-
-CACHE_SENSORS: tuple[GeocachingCacheSensorDescription, ...] = (
-    GeocachingCacheSensorDescription(
-        key="name",
-        value_fn=lambda cache: cache.name,
-    ),
-    GeocachingCacheSensorDescription(
-        key="owner",
-        value_fn=lambda cache: cache.owner.username,
-    ),
-    GeocachingCacheSensorDescription(
-        key="found",
-        value_fn=lambda cache: None
-        if cache.found_by_user is None
-        else "Yes"
-        if cache.found_by_user is True
-        else "No",
-    ),
-    GeocachingCacheSensorDescription(
-        key="found_date",
-        device_class=SensorDeviceClass.DATE,
-        value_fn=lambda cache: cache.found_date_time,
-    ),
-    GeocachingCacheSensorDescription(
-        key="favorite_points",
-        native_unit_of_measurement="points",
-        value_fn=lambda cache: cache.favorite_points,
-    ),
-    GeocachingCacheSensorDescription(
-        key="hidden_date",
-        device_class=SensorDeviceClass.DATE,
-        value_fn=lambda cache: cache.hidden_date,
-    ),
-)
-
-TRACKABLE_SENSORS: tuple[GeocachingTrackableEntityDescription, ...] = (
-    GeocachingTrackableEntityDescription(
-        key="name",
-        value_fn=lambda trackable: trackable.name,
-    ),
-    GeocachingTrackableEntityDescription(
-        key="owner",
-        value_fn=lambda trackable: trackable.owner.username,
-    ),
-    GeocachingTrackableEntityDescription(
-        key="traveled_distance",
-        native_unit_of_measurement="km",
-        value_fn=lambda trackable: None
-        if trackable.kilometers_traveled is None
-        else round(trackable.kilometers_traveled),
-    ),
-    GeocachingTrackableEntityDescription(
-        key="current_cache_code",
-        value_fn=lambda trackable: trackable.current_geocache_code,
-    ),
-    GeocachingTrackableEntityDescription(
-        key="current_cache_name",
-        value_fn=lambda trackable: trackable.current_geocache_name,
-    ),
-    GeocachingTrackableEntityDescription(
-        key="release_date",
-        device_class=SensorDeviceClass.DATE,
-        value_fn=lambda trackable: trackable.release_date,
-    ),
-)
-
-
-# pylint: disable=hass-enforce-class-module
-class GeoEntityCacheSensorEntity(GeoEntityBaseCache, SensorEntity):
-    """Representation of a cache sensor."""
-
-    entity_description: GeocachingCacheSensorDescription
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        cache: GeocachingCache,
-        description: GeocachingCacheSensorDescription,
-        category: GeocacheCategory,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator, cache, description.key, category)
-        self.entity_description = description
-
-    @property
-    def native_value(self) -> StateType | datetime.date:
-        """Return the state of the sensor."""
-        return self.entity_description.value_fn(self.cache)
-
-
-# pylint: disable=hass-enforce-class-module
-class GeoEntityTrackableSensorEntity(GeoEntityBaseTrackable, SensorEntity):
-    """Representation of a trackable sensor."""
-
-    entity_description: GeocachingTrackableEntityDescription
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        coordinator: GeocachingDataUpdateCoordinator,
-        trackable: GeocachingTrackable,
-        description: GeocachingTrackableEntityDescription,
-    ) -> None:
-        """Initialize the Geocaching sensor."""
-        super().__init__(coordinator, trackable, description.key)
-        self.entity_description = description
-
-    @property
-    def native_value(self) -> StateType | datetime.date:
-        """Return the state of the sensor."""
-        return self.entity_description.value_fn(self.trackable)
-
-
-def get_cache_entities(
-    coordinator: GeocachingDataUpdateCoordinator,
-    cache: GeocachingCache,
-    category: GeocacheCategory,
-) -> list[GeoEntityBaseCache]:
-    """Generate all entities for a single cache."""
-    entities: list[GeoEntityBaseCache] = []
-
-    # Tracker entities
-    entities.extend([GeoEntityCacheLocation(coordinator, cache, category)])
-
-    # Sensor entities
-    entities.extend(
-        [
-            GeoEntityCacheSensorEntity(coordinator, cache, description, category)
-            for description in CACHE_SENSORS
-        ]
-    )
-
-    return entities
-
-
-def get_trackable_entities(
-    coordinator: GeocachingDataUpdateCoordinator, trackable: GeocachingTrackable
-) -> list[GeoEntityBaseTrackable]:
-    """Generate all entities for a single trackable."""
-
-    entities: list[GeoEntityBaseTrackable] = []
-
-    # Tracker entities
-    entities.extend([GeoEntityTrackableLocation(coordinator, trackable)])
-
-    # Sensor entities
-    entities.extend(
-        [
-            GeoEntityTrackableSensorEntity(coordinator, trackable, description)
-            for description in TRACKABLE_SENSORS
-        ]
-    )
-
-    return entities
