+++ homeassistant/components/geocaching/sensor.py
@@ -4,21 +4,28 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from dataclasses import dataclass
+import datetime
 from typing import cast
 
-from geocachingapi.models import GeocachingStatus
+from geocachingapi.models import GeocachingCache, GeocachingStatus, GeocachingTrackable
 
-from homeassistant.components.sensor import SensorEntity, SensorEntityDescription
+from homeassistant.components.sensor import (
+    SensorDeviceClass,
+    SensorEntity,
+    SensorEntityDescription,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN, GEOCACHING_ID_SENSOR_FORMAT, GeocacheCategory
 from .coordinator import GeocachingDataUpdateCoordinator
-from .entities import get_cache_entities, get_trackable_entities
+from .device_tracker import GeoEntityCacheLocation, GeoEntityTrackableLocation
+from .entity import GeoEntityBaseCache, GeoEntityBaseTrackable
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -142,3 +149,172 @@ class GeocachingSensor(
     def native_value(self) -> str | int | None:
         """Return the state of the sensor."""
         return self.entity_description.value_fn(self.coordinator.data)
+
+
+class GeoEntityTrackableSensorEntity(GeoEntityBaseTrackable, SensorEntity):
+    """Representation of a trackable sensor."""
+
+    entity_description: GeocachingTrackableEntityDescription
+    _attr_has_entity_name = True
+
+    def __init__(
+        self,
+        coordinator: GeocachingDataUpdateCoordinator,
+        trackable: GeocachingTrackable,
+        description: GeocachingTrackableEntityDescription,
+    ) -> None:
+        """Initialize the Geocaching sensor."""
+        super().__init__(coordinator, trackable, description.key)
+        self.entity_description = description
+
+    @property
+    def native_value(self) -> StateType | datetime.date:
+        """Return the state of the sensor."""
+        return self.entity_description.value_fn(self.trackable)
+
+
+@dataclass(frozen=True, kw_only=True)
+class GeocachingTrackableEntityDescription(SensorEntityDescription):
+    """Define Sensor entity description class."""
+
+    value_fn: Callable[[GeocachingTrackable], StateType | datetime.date]
+
+
+TRACKABLE_SENSORS: tuple[GeocachingTrackableEntityDescription, ...] = (
+    GeocachingTrackableEntityDescription(
+        key="name",
+        value_fn=lambda trackable: trackable.name,
+    ),
+    GeocachingTrackableEntityDescription(
+        key="owner",
+        value_fn=lambda trackable: trackable.owner.username,
+    ),
+    GeocachingTrackableEntityDescription(
+        key="traveled_distance",
+        native_unit_of_measurement="km",
+        value_fn=lambda trackable: None
+        if trackable.kilometers_traveled is None
+        else round(trackable.kilometers_traveled),
+    ),
+    GeocachingTrackableEntityDescription(
+        key="current_cache_code",
+        value_fn=lambda trackable: trackable.current_geocache_code,
+    ),
+    GeocachingTrackableEntityDescription(
+        key="current_cache_name",
+        value_fn=lambda trackable: trackable.current_geocache_name,
+    ),
+    GeocachingTrackableEntityDescription(
+        key="release_date",
+        device_class=SensorDeviceClass.DATE,
+        value_fn=lambda trackable: trackable.release_date,
+    ),
+)
+
+
+def get_trackable_entities(
+    coordinator: GeocachingDataUpdateCoordinator, trackable: GeocachingTrackable
+) -> list[GeoEntityBaseTrackable]:
+    """Generate all entities for a single trackable."""
+
+    entities: list[GeoEntityBaseTrackable] = []
+
+    # Tracker entities
+    entities.extend([GeoEntityTrackableLocation(coordinator, trackable)])
+
+    # Sensor entities
+    entities.extend(
+        [
+            GeoEntityTrackableSensorEntity(coordinator, trackable, description)
+            for description in TRACKABLE_SENSORS
+        ]
+    )
+
+    return entities
+
+
+class GeoEntityCacheSensorEntity(GeoEntityBaseCache, SensorEntity):
+    """Representation of a cache sensor."""
+
+    entity_description: GeocachingCacheSensorDescription
+    _attr_has_entity_name = True
+
+    def __init__(
+        self,
+        coordinator: GeocachingDataUpdateCoordinator,
+        cache: GeocachingCache,
+        description: GeocachingCacheSensorDescription,
+        category: GeocacheCategory,
+    ) -> None:
+        """Initialize the Geocaching sensor."""
+        super().__init__(coordinator, cache, description.key, category)
+        self.entity_description = description
+
+    @property
+    def native_value(self) -> StateType | datetime.date:
+        """Return the state of the sensor."""
+        return self.entity_description.value_fn(self.cache)
+
+
+@dataclass(frozen=True, kw_only=True)
+class GeocachingCacheSensorDescription(SensorEntityDescription):
+    """Define Sensor entity description class."""
+
+    value_fn: Callable[[GeocachingCache], StateType | datetime.date]
+
+
+CACHE_SENSORS: tuple[GeocachingCacheSensorDescription, ...] = (
+    GeocachingCacheSensorDescription(
+        key="name",
+        value_fn=lambda cache: cache.name,
+    ),
+    GeocachingCacheSensorDescription(
+        key="owner",
+        value_fn=lambda cache: cache.owner.username,
+    ),
+    GeocachingCacheSensorDescription(
+        key="found",
+        value_fn=lambda cache: None
+        if cache.found_by_user is None
+        else "Yes"
+        if cache.found_by_user is True
+        else "No",
+    ),
+    GeocachingCacheSensorDescription(
+        key="found_date",
+        device_class=SensorDeviceClass.DATE,
+        value_fn=lambda cache: cache.found_date_time,
+    ),
+    GeocachingCacheSensorDescription(
+        key="favorite_points",
+        native_unit_of_measurement="points",
+        value_fn=lambda cache: cache.favorite_points,
+    ),
+    GeocachingCacheSensorDescription(
+        key="hidden_date",
+        device_class=SensorDeviceClass.DATE,
+        value_fn=lambda cache: cache.hidden_date,
+    ),
+)
+
+
+def get_cache_entities(
+    coordinator: GeocachingDataUpdateCoordinator,
+    cache: GeocachingCache,
+    category: GeocacheCategory,
+) -> list[GeoEntityBaseCache]:
+    """Generate all entities for a single cache."""
+    entities: list[GeoEntityBaseCache] = []
+
+    # Tracker entities
+    entities.extend([GeoEntityCacheLocation(coordinator, cache, category)])
+
+    # Sensor entities
+    entities.extend(
+        [
+            GeoEntityCacheSensorEntity(coordinator, cache, description, category)
+            for description in CACHE_SENSORS
+        ]
+    )
+
+    return entities
